SwipeToDeleteCallback 클래스를 분리하여 책임 분리 및 가독성 개선
