+++ presentation/src/main/java/com/example/presentation/home/HomeFragment.kt
@@ -29,6 +29,7 @@ import com.example.presentation.adapter.AdapterTodo
 import com.example.presentation.databinding.FragmentHomeBinding
 import com.example.presentation.dpToPx
 import com.example.presentation.dpToPxF
+import com.example.presentation.helper.SwipeToDeleteCallback
 import com.example.presentation.showCustomToast
 import com.example.presentation.view.chip.CustomChipView
 import com.example.presentation.view.tab.SlidingTabToggleView
@@ -173,104 +174,11 @@ class HomeFragment : Fragment() {
     }
 
     private fun setupSwipeToDelete() {
-        val swipeCallback = object : ItemTouchHelper.SimpleCallback(0, ItemTouchHelper.LEFT) {
-            override fun onMove(
-                rv: RecyclerView,
-                vh: RecyclerView.ViewHolder,
-                target: RecyclerView.ViewHolder
-            ) = false
-
-            override fun onSwiped(vh: RecyclerView.ViewHolder, direction: Int) {
-                val position = vh.adapterPosition
-                showDeleteConfirmDialog(position)
-            }
-
-            override fun onChildDraw(
-                c: Canvas,
-                recyclerView: RecyclerView,
-                viewHolder: RecyclerView.ViewHolder,
-                dX: Float,
-                dY: Float,
-                actionState: Int,
-                isCurrentlyActive: Boolean
-            ) {
-                val itemView = viewHolder.itemView
-                val context = recyclerView.context
-
-                if (dX < 0) {
-                    val cornerRadius = 8.dpToPxF(context)
-                    val marginLeft = 16.dpToPx(context)
-                    val verticalMargin = 8.dpToPx(context)
-
-                    val top = itemView.top + verticalMargin
-                    val bottom = itemView.bottom - verticalMargin
-
-                    val isFullySlid = dX <= -itemView.width.toFloat()
-                    val backgroundLeft =
-                        itemView.right.toFloat() + dX + if (isFullySlid) marginLeft.toFloat() else 0f
-                    val backgroundRight = itemView.right.toFloat()
-
-                    val path = Path().apply {
-                        addRoundRect(
-                            backgroundLeft,
-                            top.toFloat(),
-                            backgroundRight,
-                            bottom.toFloat(),
-                            floatArrayOf(
-                                0f,
-                                0f,
-                                cornerRadius,
-                                cornerRadius,
-                                cornerRadius,
-                                cornerRadius,
-                                0f,
-                                0f
-                            ),
-                            Path.Direction.CW
-                        )
-                    }
-
-                    val backgroundPaint = Paint().apply {
-                        color = ContextCompat.getColor(context, R.color.secondary)
-                        isAntiAlias = true
-                    }
-
-                    c.drawPath(path, backgroundPaint)
-
-                    val text = context.getString(R.string.swipe_delete_hint)
-                    val textPaint = Paint().apply {
-                        color = Color.RED
-                        textSize = 12.dpToPxF(context)
-                        isAntiAlias = true
-                        textAlign = Paint.Align.CENTER
-                        typeface = ResourcesCompat.getFont(context, R.font.notosanskrregular)
-                    }
-
-                    val textX = (backgroundLeft + backgroundRight) / 2f
-                    val textY =
-                        itemView.top + (itemView.height / 2f - (textPaint.descent() + textPaint.ascent()) / 2)
-
-                    c.save()
-                    c.clipRect(backgroundLeft, top.toFloat(), backgroundRight, bottom.toFloat())
-                    c.drawText(text, textX, textY, textPaint)
-                    c.restore()
-                }
-
-                super.onChildDraw(
-                    c,
-                    recyclerView,
-                    viewHolder,
-                    dX,
-                    dY,
-                    actionState,
-                    isCurrentlyActive
-                )
-            }
-
+        val callback = SwipeToDeleteCallback(requireContext()) { position ->
+            showDeleteConfirmDialog(position)
         }
-        ItemTouchHelper(swipeCallback).attachToRecyclerView(binding.recyclerView)
+        ItemTouchHelper(callback).attachToRecyclerView(binding.recyclerView)
     }
-
 
     private fun showDeleteConfirmDialog(position: Int) {
         CustomAlertDialog.Builder(requireContext())
